+++ src/agedi/diffusion/__init__.py
@@ -1,5 +1,24 @@
 from .agedi import Agedi, ForcefieldGuidanceConfig
 from .diffusion import Diffusion
+from .samplers import (
+    EulerMaruyamaSampler,
+    ForcefieldCorrectorSampler,
+    HeunODESampler,
+    HeunSampler,
+    PredictorCorrectorSampler,
+    ProbabilityFlowODESampler,
+    Sampler,
+)
 
-__all__ = ["Agedi", "Diffusion", "ForcefieldGuidanceConfig"]
-
+__all__ = [
+    "Agedi",
+    "Diffusion",
+    "ForcefieldGuidanceConfig",
+    "Sampler",
+    "EulerMaruyamaSampler",
+    "PredictorCorrectorSampler",
+    "HeunSampler",
+    "ProbabilityFlowODESampler",
+    "HeunODESampler",
+    "ForcefieldCorrectorSampler",
+]